backend/requirements.txt
Django==4.2.11 
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
channels==4.0.0
celery==5.3.6
redis==5.0.3
channels-redis==4.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-celery-beat==2.5.0  # Add this line
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-cors-headers==4.3.1